recall-server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psutil
Pillow
prometheus-client
slowapi
redis
structlog
